+++ helpers/index.js
@@ -1,0 +1,3 @@
+const uploadAvatars = require('./uploadAvatars');
+
+module.exports = { uploadAvatars };
